Project#1-CustomerSupportAgent/Codes/requirements.txt
# --- Core AI & LLM Integration ---
google-genai==1.57.0
langchain==1.2.3
langchain-google-genai==4.1.3
langchain-community==0.4.1

# --- Vector Database & Data Handling ---
faiss-cpu==1.13.2
pandas==2.2.3
numpy==2.1.0
openpyxl==3.1.5

# --- UI & Visualization ---
streamlit==1.41.0
matplotlib==3.10.0